recovery: Kill dead updater code.
The vendor-specific updater extensions (BootUpdateFn and DecryptFn) are
only meaningful for non-A/B updater. They shouldn't be built (nor linked
into the "librecovery_updater_msm" module) on A/B devices.

This CL removes the updater related functions, and renames the module to
libgptutils which is used by boot control HAL.

Bug: 33607778
Test: m checkbuild
Test: Apply an update on top of the new system.
Change-Id: Ifad981602a426bd6560a2a3142f49110522f9c88