Update patch set 1
Patch Set 1:

> Are these rules still needed in audioserver.te or can they be
 > removed from there in this CL?

Not in this CL. Currently non-binderized configuration is still the default, and binderized is a special build, so we need to have the same rules in both files.

But you are right that some duplicating rules from audioserver.te may be removed later after the binderized configuration becomes the default.

Patch-set: 1