Update patch set 2
Patch Set 2:

So in 7.1.1, there's this commit:

commit 30a716157b748258e5461fe783298ba7c740835f
Author: Sunny Goyal <sunnygoyal@google.com>
Date:   Tue Apr 12 17:57:41 2016 -0700

    Removing WallpaperPicker from Launcher3

I ported this change from there, and it worked, but I hadn't actually verified that the problem occurred on master. Maybe that commit is going to show up in master later? Not really clear, since master does have the WallpaperPicker repository too.

Patch-set: 2