Update patch set 2
Patch Set 2:

> It doesn't look like 'Removing WallpaperPicker from Launcher3' is
 > in master yet (thus not reproducing), just in the repository via
 > android-7.1.1 tags. Anyway, I'm confused now :P.

The nyc build should be fine. In nyc-mr1, Launcher3 does not include wallpaper app and I dont think we are tracking the wallpaper app. 
The Launcher3 in mr1 simply starts the default wallpaper picker app on the system. But since there is no wallpaper app in AOSP build, it will end up crashing. We are currently trying to resolve this.

Patch-set: 2