Update patch set 2
Patch Set 2:

The previous max r/w of 16834 came from a hardcoded max which was removed in the 3.3 kernel. The values here come from the maximum amount of contiguous DMAable memory that the kernel can allocate. I'm not sure where the values are set but I think it's somewhere in the kernel allocator layer.

Patch-set: 2