Update patch set 2
Patch Set 2:

> Doesn't this break power hal for sailfish?

No, it shouldn't. device-common.mk is loaded by both marlin and sailfish, so we'll set ro.hardware.power=marlin for both.

> "power.msm8996" would have been more accurate anyway.

Perhaps, but this is a marlin/sailfish specific power hal, if only because it's in the marlin device folder. If it was under hardware/qcom so that you'd only have a single copy, then yes, msm8996 may be more correct.

Patch-set: 2