build toybox_static for modprobe in recovery mode
toybox is needed for modprobe in recovery mode (for Treble). This change adds
toybox_static to PRODUCT_PACKAGES only for sailfish/marlin for now. The size of
statically linked toybox is around 1.2MB and this fits well in the boot
partition of sailfish/marlin whose size is set to 32MB.

Bug: 35633646
Test: Build and tested it works in recovery with sailfish
Change-Id: I26350f638e6a9dab73980c7b15a4278e9f1c080d