marlin: copy vndk-sp libraries to /vendor/lib/vndk-sp
Note: this shouldn't be permanent solution. Anything that is installed
to /vendor partition must be compiled against the vndk build of the
system libraries, which might be different than the core build of the
libraries. However, since the number of such libraries are so many and
thus it takes considerable time to build them as vndk. So, in the mean
time, we just build vendor libraries against the core build as before
and copy the core builds to the vendor partition.

Test: the listed files are in /vendor/lib(64?)/vndk-sp and their
contents are the same as those in /system/lib(64?).

(cherry picked from commit 24b85fdee514e517ca635db83dfbeff56c92b242)

Change-Id: I6ca53149705e3d72c43c9770210c890533ad1a18