DO NOT MERGE: genfscon vendor extensions
Add a new attribute in system/public/attributes for vendor/genfs_contexts
to use. It still can work after using GSI, because all types are defined
in vendor and only attributes are in system (which are allowed to be absent).

plat_sepolic.cil (allowed to be absent, when GSI is used):
   - (typeattribute sysfs_oem_sensor_27_0)
   - (allow oem_process sysfs_oem_sensor_27_0 (file (write lock append map open)))

nonplat_sepolicy.cil (self-contained):
   - (typeattribute sysfs_oem_sensor_27_0)
   - (typeattributeset sysfs_oem_sensor_27_0 (sysfs_oem_sensor))
   - (type sysfs_oem_sensor)
   - (roletype object_r sysfs_oem_sensor)
   - (allow vnd_process sysfs_oem_sensor (file (write lock append map open)))
   - (genfscon sysfs /devices/virtual/oem_sensor (u object_r sysfs_oem_sensor ((s0) (s0))))

Bug: 64504042
Test: None
Change-Id: I53f08c8e292475071936579f985f237858062de4