preopt: Make GoogleCamera compile speed-profile, saves ~17MB ram
GoogleCamera is currently compiled as speed, but we now have a
GoogleCamera.perf.txt in the source tree so it should be always
preopted as perf.

The PinnerService currently pins the entire GoogleCamera ODEX file
into memory always and forever.

At the time of writing:
* GoogleCamera.odex size (speed) is 18.1MB
* GoogleCamera.odex size (speed-profile) is ~500KB

This simple change saves about 17.5MB of memory for the out-of-box
experience.

Change-Id: I47e3b1bde4550407069324dc943a896f113da22f