Remove Runtime APEX internal libs from the vndk-sp list.
libdexfile_support is the only direct dependency for Renderscript, and it is
also intended to be on the caller side.

Even if the APEX won't be updated on VNDK-lite devices like marlin, the APEX
internal libraries still need to be solely under /apex/com.android.runtime,
to make them get loaded in the correct namespace.

Test: Flash and boot on marlin, AOSP and internal
Test: atest CtsRenderscriptTestCases on marlin, AOSP and internal
Bug: 124293228
Change-Id: Ie3cae4b636253c8ecca9c01ab06c2513112f9771